+++ sdk/js/src/util/index.ts
@@ -1,3 +1,2 @@
 export * from "./buffer-util.js";
 export * from "./env-util.js";
-export * from "./url-util.js";